Functionality for chat, version 1.
This is a very belated push. I wanted to do a different approach but couldn't get it to work. So I had to redo a lot of things, so none of the code felt suitable to push.

Summary:
* Added three layouts.
- activity_chat.
- fragment_chat.
- chat_list_item.
* Added the ChatViewModel.
* Added the adapter (ChatAdapter) and fragment (ChatFragment) for bindings / mapping(?).
* Added ChatClient, to spread out some of the code from FirebaseUtils.
* Added models needed for the Firestore data.
- MessageData.
- ReactionData.
- SessionData.

Details:
In MainActivity you call the layout for activity_chat. Which contains a FragmentContainerView. Each message is a fragment. This code / structure is taken from the My Places application. Though it's different because I'm using Kotlin StateFlow, and the Firebase Snapshot listener.
A lot of refactoring, comments, and cleanup are needed. Though this works now, the user can send a message which is displayed to themselves and others.

Next:
* Polish the layout. Needs to be reactive and have the list of participating users.
* Set up the chat activity so we can call it from the friends screen.
* Extra: Reactions and editing.